+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:example/pages/demo/appbar.dart';
 import 'package:example/pages/demo/button.dart';
 import 'package:example/pages/demo/cascader.dart';
+import 'package:example/pages/demo/checkbutton.dart';
 import 'package:example/pages/demo/dialog.dart';
 import 'package:example/pages/demo/empty.dart';
 import 'package:example/pages/demo/error.dart';
@@ -10,6 +11,7 @@ import 'package:example/pages/demo/image.dart';
 import 'package:example/pages/demo/listview.dart';
 import 'package:example/pages/demo/loading.dart';
 import 'package:example/pages/demo/native.dart';
+import 'package:example/pages/demo/qrcode.dart';
 import 'package:example/pages/demo/searchbar.dart';
 import 'package:example/pages/demo/tabpageview.dart';
 import 'package:example/pages/demo/textfield.dart';
@@ -54,6 +56,7 @@ class MyRouter {
   static final appbarPage = '/demo/appbar';
   static final searchbarPage = '/demo/searchbar';
   static final buttonPage = '/demo/button';
+  static final checkButtonPage = '/demo/checkbutton';
   static final cascaderPage = '/demo/cascader';
   static final textFieldPage = '/demo/textfield';
   static final dialogPage = '/demo/dialog';
@@ -66,6 +69,8 @@ class MyRouter {
   static final videoPage = '/demo/video';
   static final tabViewPage = '/demo/tabview';
   static final listViewPage = '/demo/listview';
+  static final qrCodePage = '/demo/qrcode';
+
 
   //
   // 路由代码
@@ -78,6 +83,7 @@ class MyRouter {
     appbarPage: (ctx) => AppBarPage(),
     searchbarPage: (ctx) => SearchBarPage(),
     buttonPage: (ctx) => ButtonPage(),
+    checkButtonPage: (ctx) => CheckButtonPage(),
     cascaderPage: (ctx) => CascaderPage(),
     textFieldPage: (ctx) => TextFieldPage(),
     dialogPage: (ctx) => DialogPage(),
@@ -90,6 +96,7 @@ class MyRouter {
     videoPage: (ctx) => VideoPage(),
     tabViewPage: (ctx) => TabPageViewPage(),
     listViewPage: (ctx) => ListViewPage(),
+    qrCodePage: (ctx) => QRCodePage(),
   };
 }