+++ example/lib/pages/demo/cascader.dart
@@ -40,8 +40,8 @@ class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin
       dynamic  arguments = ModalRoute.of(context)!.settings.arguments;
       print("路由传递过来的参数：" + arguments.toString());
       if (arguments != null) {
-        isRouteFromFlutter = arguments["flutter"];
-        if(Config.isWeb || isRouteFromFlutter){
+        isRouteFlutter = arguments["flutter"];
+        if(!Config.isNative || isRouteFlutter){
           //加载数据
           _loadData(0,0);
         }