+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## [0.0.8] - 2021/07/10.
+* add YmCheckButton and YmRadioButton Sample
+* add YmAppBar right button sample
+* add YmQrCode add qr code scan Sample
+
 ## [0.0.7] - 2021/06/30.
 * add YmCheckButton
 * add YmRadioButton